feat : [SF] add home component four and mock data recommendation

+++ lib/app/mock_data/model/product.dart
@@ -1,9 +1,9 @@
 class Product {
   late String image;
   late String name;
-  late int price;
+  late double price;
 
-  Product(String image, String name, int price){
+  Product(String image, String name, double price){
     this.image = image;
     this.name = name;
     this.price = price;
